allow FASTQ to be duplicated if "repeated" in notes
This addresses #104 by @bblarsen-sci

The solution here is that in order to repeat a FASTQ, the *notes* column in `barcode_runs.csv` for all rows with the repeated entry must have the word "repeated".

This ensures that in general we still have a check that FASTQs are not being re-used, since that could often indicate an error.

But in cases where there is a good reason to re-use one, you will have a note in the *notes* column, like "this FASTQ repeated in replicates XXX and YYY because ZZZ"